Revert "Rename folders"
This reverts commit 31b7ff949d54bce444716b0cf1f276ae3e21887a.
