+++ packages/doenetml-worker-rust/lib-doenetml-core/src/core.rs
@@ -3,7 +3,7 @@ pub mod components;
 pub mod dast;
 pub mod dependency;
 mod doenet_ml_core;
-pub mod generic_state_var;
 pub mod state;
+pub mod state_var_interfaces;
 
 pub use self::doenet_ml_core::*;
